+++ components/media/geometry/sources/core/geometry_mesh_check.cpp
@@ -45,15 +45,15 @@ class Log
 
 struct CheckContext
 {
-  const VertexStream&        stream;        //вершинный массив
-  const VertexWeightStream&  weights;       //веса
-  size_t                     vertex_buffer_index; //номер вершинного буфера
-  size_t                     stream_index;  //номер вершинного массива
-  size_t                     weights_count; //количество весов
-  size_t                     vertex_index;  //номер вершины
+  const VertexStream&        stream;              //вершинный массив
+  const VertexWeightStream&  weights;             //веса
+  unsigned int               vertex_buffer_index; //номер вершинного буфера
+  unsigned int               stream_index;        //номер вершинного массива
+  unsigned int               weights_count;       //количество весов
+  unsigned int               vertex_index;        //номер вершины
   const VertexAttribute*     attribute;
   
-  CheckContext (const VertexBuffer& vb, size_t in_vertex_buffer_index, size_t in_stream_index) :
+  CheckContext (const VertexBuffer& vb, unsigned int in_vertex_buffer_index, unsigned int in_stream_index) :
     stream (vb.Stream (in_stream_index)), weights (vb.Weights ()), vertex_buffer_index (in_vertex_buffer_index), stream_index (in_stream_index),
     weights_count (weights.Size ()), vertex_index (0), attribute (0) {}
 };
@@ -82,7 +82,7 @@ inline void check_attribute (Log& log, CheckContext& context, const VertexInflue
   const VertexWeight* weight      = context.weights.Data () + influence.first_weight;
   float               weights_sum = 0.0f;
   
-  for (size_t i=0; i<influence.weights_count; i++, weight++)
+  for (unsigned int i=0; i<influence.weights_count; i++, weight++)
     weights_sum += weight->joint_weight;
     
   if (fabs (weights_sum - 1.0f) > EPS)
@@ -95,7 +95,7 @@ inline void check_attribute (Log& log, CheckContext& context, const math::vector
 {
   static const char* component_name [] = {"x", "y", "z", "w"};
 
-  for (size_t i=0; i<Size; i++)
+  for (unsigned int i=0; i<Size; i++)
   {
     #ifdef _MSC_VER
       #define isfinite _finite
@@ -113,10 +113,10 @@ inline void check_attribute (Log& log, CheckContext& context, const math::vector
 template <class T>
 void check_stream (Log& log, CheckContext& context, const char* iter)
 {
-  size_t verts_count = context.stream.Size (),
-         stride      = context.stream.VertexSize ();
+  unsigned int verts_count = context.stream.Size (),
+               stride      = context.stream.VertexSize ();
 
-  for (size_t i=0; i<verts_count; i++, iter += stride)
+  for (unsigned int i=0; i<verts_count; i++, iter += stride)
   {
     context.vertex_index = i;
     check_attribute (log, context, *reinterpret_cast<const T*> (iter));
@@ -127,7 +127,7 @@ void check_stream (Log& log, CheckContext& context)
 {  
   const VertexFormat& format = context.stream.Format ();
     
-  for (size_t i=0, count=format.AttributesCount (); i<count; i++)
+  for (unsigned int i=0, count=format.AttributesCount (); i<count; i++)
   {
     const VertexAttribute& attribute = format.Attribute (i);
 
@@ -178,8 +178,8 @@ void check_stream (Log& log, CheckContext& context)
       continue;
     }
 
-    size_t vertex_size = context.stream.VertexSize (),
-           type_size   = get_type_size (attribute.type);
+    unsigned int vertex_size = context.stream.VertexSize (),
+                 type_size   = get_type_size (attribute.type);
     
     if (attribute.offset > vertex_size || vertex_size - attribute.offset < type_size)
     {
@@ -214,9 +214,9 @@ void check_stream (Log& log, CheckContext& context)
 }
 
 template <class T>
-void check_indices (size_t count, const T* index, size_t verts_count, size_t base_vertex, size_t primitive_index, Log& log)
+void check_indices (unsigned int count, const T* index, unsigned int verts_count, unsigned int base_vertex, unsigned int primitive_index, Log& log)
 {
-  for (size_t j=0; j<count; j++, index++)
+  for (unsigned int j=0; j<count; j++, index++)
     if (*index + base_vertex >= verts_count)
     {
       if (base_vertex)
@@ -255,7 +255,7 @@ bool check (const Mesh& mesh, uint32_t joints_count, const xtl::function<void (c
     
     bool has_indices = mesh.IndexBuffer ().Size () != 0;
 
-    for (size_t i=0, count=mesh.PrimitivesCount (); i<count; i++)
+    for (unsigned int i=0, count=mesh.PrimitivesCount (); i<count; i++)
     {                  
       const Primitive& primitive = mesh.Primitive (i);
       
@@ -267,8 +267,8 @@ bool check (const Mesh& mesh, uint32_t joints_count, const xtl::function<void (c
 
       const VertexBuffer& vertex_buffer = mesh.VertexBuffer (primitive.vertex_buffer);
 
-      size_t      max_index      = has_indices ? index_buffer.Size () : vertex_buffer.VerticesCount ();    
-      const char* max_index_name = has_indices ? "indices_count" : "vertices_count";      
+      unsigned int max_index      = has_indices ? index_buffer.Size () : vertex_buffer.VerticesCount ();
+      const char*  max_index_name = has_indices ? "indices_count" : "vertices_count";
       
       switch (primitive.type)
       {
@@ -303,7 +303,7 @@ bool check (const Mesh& mesh, uint32_t joints_count, const xtl::function<void (c
       
         //проверка корректности индексов
         
-      size_t verts_count = vertex_buffer.VerticesCount ();        
+      unsigned int verts_count = vertex_buffer.VerticesCount ();
       
       switch (index_buffer.DataType ())
       {
@@ -324,13 +324,13 @@ bool check (const Mesh& mesh, uint32_t joints_count, const xtl::function<void (c
         
       //проверка вершинных буферов
       
-    for (size_t i=0, count=mesh.VertexBuffersCount (); i<count; i++)
+    for (unsigned int i=0, count=mesh.VertexBuffersCount (); i<count; i++)
     {
         //проверка вершинных массивов
 
       const VertexBuffer& vertex_buffer = mesh.VertexBuffer (i);
 
-      for (size_t j=0, count=vertex_buffer.StreamsCount (); j<count; j++)
+      for (unsigned int j=0, count=vertex_buffer.StreamsCount (); j<count; j++)
       {
         CheckContext check_context (vertex_buffer, i, j);
         
@@ -341,7 +341,7 @@ bool check (const Mesh& mesh, uint32_t joints_count, const xtl::function<void (c
         
       const VertexWeight* weight = vertex_buffer.Weights ().Data ();
         
-      for (size_t j=0, count=vertex_buffer.Weights ().Size (); j<count; j++, weight++)
+      for (unsigned int j=0, count=vertex_buffer.Weights ().Size (); j<count; j++, weight++)
       {
         if (weight->joint_index >= joints_count)
           log.Error ("vertex_buffer[%u].weight[%u].joint_index=%u >= joints_count=%u", i, j, weight->joint_index, joints_count);
